+++ app/modules/common/factories/UserProfile.factory.js
@@ -1,260 +1,399 @@
 angular.module('common')
-.factory('UserProfile', function(API,LocaleService,$q,$timeout) {
+.factory('UserProfile', function(API, APIError, LocaleService, Timezones, $filter, $q, $timeout) {
 
-        var self = {
-            getProfile: function(userCredentials) {
-                var defer = $q.defer(),
-                    userProfile = {};
+    const errorName = 'userProfileFactory.'
 
-                API.cui.getPerson(userCredentials)
-                .then(function(res) {
-                    if (!res.addresses) {
-                        // If the person has no addresses set we need to initialize it as an array
-                        // to follow the object structure
-                        res.addresses = [{}];
-                    }
-                    if (!res.addresses[0].streets) {
-                        res.addresses[0].streets = [];
-                    }
-                    userProfile.user = {};
-                    userProfile.tempUser = {};
-                    angular.copy(res, userProfile.user);
-                    angular.copy(res, userProfile.tempUser);
-                    return API.cui.getSecurityQuestionAccount(userCredentials);
-                })
-                .then(function(res) {
-                    userProfile.userSecurityQuestions = res;
-                    userProfile.tempUserSecurityQuestions = angular.copy(userProfile.userSecurityQuestions.questions);
-                    return API.cui.getSecurityQuestions();
-                })
-                .then(function(res) {
-                    userProfile.allSecurityQuestions = res;
-                    userProfile.allSecurityQuestionsDup = angular.copy(res);
-                    userProfile.allSecurityQuestions.splice(0,1);
+    const UserProfile = {
 
-                    // Splits questions to use between both dropdowns
-                    var numberOfQuestions = userProfile.allSecurityQuestions.length,
-                        numberOfQuestionsFloor = Math.floor(numberOfQuestions/3);
+        initUser: function(userId) {
+            let defer = $q.defer()
+            let user = {}
 
-                    //Allocating options to three questions
-                    userProfile.allChallengeQuestions0 = userProfile.allSecurityQuestions.splice(0,numberOfQuestionsFloor);
-                    userProfile.allChallengeQuestions1 = userProfile.allSecurityQuestions.splice(0,numberOfQuestionsFloor);
-                    userProfile.allChallengeQuestions2 = userProfile.allSecurityQuestions.splice(0,numberOfQuestionsFloor);
-
-                    self.selectTextsForQuestions(userProfile);
-                    return API.cui.getOrganization({organizationId:userProfile.user.organization.id});
-                    })
-                    .then(function(res) {
-                        userProfile.organization = res;
-                        return API.cui.getPasswordPolicy({policyId: res.passwordPolicy.id});
-                    })
-                    .then(function(res) {
-                        userProfile.passwordRules = res.rules;
-                        defer.resolve( userProfile );
-                    })
-                    .fail(function(err) {
-                        console.error('UserProfile.factory error: ', err);
-                        defer.reject(err);
-                    });
-
-                return defer.promise;
-            },
-
-            selectTextsForQuestions: function(userProfile) {
-                userProfile.challengeQuestionsTexts = [];
-                angular.forEach(userProfile.userSecurityQuestions.questions, function(userQuestion) {
-                    var question = _.find(userProfile.allSecurityQuestionsDup, function(question) {
-                        return question.id === userQuestion.question.id;
-                    });
-                    this.push(question.question[0].text);
-                }, userProfile.challengeQuestionsTexts);
-            },
-
-            resetTempUser: function(userProfile) {
-                if (!angular.equals(userProfile.tempUser,userProfile.user)) {
-                    angular.copy(userProfile.user,userProfile.tempUser);
+            API.cui.getPerson({ personId: userId })
+            .done(res => {
+                // If the person object has no addresses we need to initialize it
+                if (!res.addresses) {
+                    res.addresses = [{streets: []}];
                 }
-            },
+                user.user = Object.assign({}, res)
+                user.tempUser = Object.assign({}, res)
+                defer.resolve(user)
+            })
+            .fail(err => {
+                console.error('Failed getting user information', err)
+                APIError.onFor(errorName + 'initUser')
+                $timeout(() => {
+                    APIError.offFor(errorName + 'initUser')
+                }, 5000)
+                defer.reject(err)
+            })
+            return defer.promise
+        },
 
-            getPersonPasswordAccount: function(userProfile) {
-                return {
-                    version: '1',
-                    username: userProfile.user.username,
-                    currentPassword: userProfile.userPasswordAccount.currentPassword,
-                    password: userProfile.userPasswordAccount.password,
-                    passwordPolicy: userProfile.organization.passwordPolicy,
-                    authenticationPolicy: userProfile.organization.authenticationPolicy
-                };
-            },
+        initSecurityQuestions: function(userId) {
+            let defer = $q.defer()
+            let securityQuestions = {
+                userSecurityQuestions: {},
+                tempUserSecurityQuestions: {},
+                allSecurityQuestions: [],
+                allSecurityQuestionsDup: []
+            }
 
-            injectUI: function(userProfile, $scope, personId) {
-                let userId
-                personId ? userId = personId : userId = API.getUser()
+            $q.all([
+                API.cui.getSecurityQuestionAccount({ personId: userId }), 
+                API.cui.getSecurityQuestions()
+            ])
+            .then(res => {
+                angular.copy(res[0], securityQuestions.userSecurityQuestions)
+                angular.copy(res[0], securityQuestions.tempUserSecurityQuestions)
+                angular.copy(res[1], securityQuestions.allSecurityQuestions) 
+                angular.copy(res[1], securityQuestions.allSecurityQuestionsDup)
 
-                userProfile.toggleAllOff = function() {
-                    angular.forEach(userProfile.toggleOffFunctions, function(toggleOff) {
-                        toggleOff();
-                    });
-                    self.resetTempUser(userProfile);
-                };
+                securityQuestions.allSecurityQuestions.splice(0, 1)
 
-                userProfile.resetTempObject = function(master, temp) {
-                    // Used to reset the temp object to the original when a user cancels their edit changes
-                    if (!angular.equals(master, temp)) {
-                        angular.copy(master, temp);
+                let numberOfQuestions = securityQuestions.allSecurityQuestions.length
+                let numberOfQuestionsFloor = Math.floor(numberOfQuestions/2)
+
+                securityQuestions.allChallengeQuestions0 = securityQuestions.allSecurityQuestions.slice(0, numberOfQuestionsFloor)
+                securityQuestions.allChallengeQuestions1 = securityQuestions.allSecurityQuestions.slice(numberOfQuestionsFloor)
+
+                securityQuestions.challengeQuestionsTexts = UserProfile.selectTextsForQuestions(securityQuestions)
+
+                defer.resolve(securityQuestions)
+            })
+            .catch(err => {
+                console.error('Failed getting security question data', err)
+                APIError.onFor(errorName + 'initSecurityQuestions')
+                $timeout(() => {
+                    APIError.offFor(errorName + 'initSecurityQuestions')
+                }, 5000)
+                defer.reject(err)
+            })
+            return defer.promise
+        },
+
+        selectTextsForQuestions: function(securityQuestions) {
+            let challengeQuestionsTexts = []
+
+            angular.forEach(securityQuestions.userSecurityQuestions.questions, (userQuestion) => {
+                let question = _.find(securityQuestions.allSecurityQuestionsDup, (question) => {
+                    return question.id === userQuestion.question.id
+                })
+                challengeQuestionsTexts.push($filter('cuiI18n')(question.question))
+            })
+            return challengeQuestionsTexts
+        },
+
+        initPasswordPolicy: function(organizationId) {
+            let defer = $q.defer()
+            let passwordPolicy = {}
+
+            API.cui.getOrganization({ organizationId: organizationId })
+            .then(res => {
+                passwordPolicy.organization = res
+                return API.cui.getPasswordPolicy({policyId: res.passwordPolicy.id})
+            })
+            .then(res => {
+                passwordPolicy.passwordRules = res.rules
+                res.rules.forEach(rule => {
+                    if (rule.type === 'history') {
+                        passwordPolicy.numberOfPasswords = rule.numberOfPasswords
                     }
-                };
+                })
+                defer.resolve(passwordPolicy)
+            })
+            .fail(err => {
+                console.error('Failed getting password policy data', err)
+                APIError.onFor(errorName + 'initPasswordPolicy')
+                $timeout(() => {
+                    APIError.offFor(errorName + 'initPasswordPolicy')
+                }, 5000)
+                defer.reject(err)
+            })
+            return defer.promise
+        },
 
-                userProfile.resetPasswordFields = function() {
-                    // Used to set the password fields to empty when a user clicks cancel during password edit
-                    userProfile.userPasswordAccount = {
-                        currentPassword: '',
-                        password: ''
-                    };
-                    userProfile.passwordRe = '';
-                };
+        initRegisteredDate: function(userId) {
+            const defer = $q.defer()
 
-                userProfile.checkIfRepeatedSecurityAnswer = function(securityQuestions, formObject) {
-                    securityQuestions.forEach(function(secQuestion, i) {
-                        var securityAnswerRepeatedIndex = _.findIndex(securityQuestions, function(secQuestionToCompareTo,z) {
-                            return z !== i && secQuestion.answer && secQuestionToCompareTo.answer && secQuestion.answer.toUpperCase() === secQuestionToCompareTo.answer.toUpperCase();
-                        });
-                        if (securityAnswerRepeatedIndex > -1) {
-                            if (formObject['answer' + securityAnswerRepeatedIndex]) {
-                                formObject['answer' + securityAnswerRepeatedIndex].$setValidity('securityAnswerRepeated', false);
-                            }
-                            if(formObject['answer' + i]) {
-                                formObject['answer' + i].$setValidity('securityAnswerRepeated', false);
-                            }
+            API.cui.getPersonStatusHistory({qs : [
+                ['userId', userId], 
+                ['sortBy', '+creation']
+            ]})
+            .then(res => {
+                res.forEach(status => {
+                    if (status.status === 'ACTIVE') {
+                        defer.resolve(status.creation)             
+                    }
+                })
+            })
+            .fail(error => {
+                console.error('initRegisteredDate: There was an issue retrieving the registered date.')
+                APIError.onFor(errorName + 'initRegisteredDate')
+                $timeout(() => {
+                    APIError.offFor(errorName + 'initRegisteredDate')
+                }, 5000)
+                defer.reject(error)
+            })
+
+            return defer.promise
+        },
+
+        initUserProfile: function(userId, organizationId) {
+            let defer = $q.defer()
+            let profile = {}
+            let callsCompleted = 0
+            const callsToComplete = 4
+
+            UserProfile.initUser(userId)
+            .then(res => {
+                angular.merge(profile, res)
+            })
+            .finally(() => {
+                callsCompleted += 1
+                if (callsCompleted === callsToComplete) defer.resolve(profile)
+            })
+
+            UserProfile.initSecurityQuestions(userId)
+            .then(res => {
+                angular.merge(profile, res)
+            })
+            .finally(() => {
+                callsCompleted += 1
+                if (callsCompleted === callsToComplete) defer.resolve(profile)
+            })
+
+            UserProfile.initPasswordPolicy(organizationId)
+            .then(res => {
+                angular.merge(profile, res)
+            })
+            .finally(() => {
+                callsCompleted += 1
+                if (callsCompleted === callsToComplete) defer.resolve(profile)
+            })
+
+            UserProfile.initRegisteredDate(userId)
+            .then(res => {
+                profile['registeredDate'] = res
+            })
+            .finally(() => {
+                callsCompleted += 1
+                if (callsCompleted === callsToComplete) defer.resolve(profile)
+            })
+
+            return defer.promise
+        },
+
+        buildPersonPasswordAccount: function(user, passwordAccount, organization) {
+            return {
+                version: '1',
+                username: user.username,
+                currentPassword: passwordAccount.currentPassword,
+                password: passwordAccount.password,
+                passwordPolicy: organization.passwordPolicy,
+                authenticationPolicy: organization.authenticationPolicy
+            }
+        },
+
+        injectUI: function(profile, $scope, personId) {
+            let userId
+
+            personId
+                ? userId = personId
+                : userId = API.getUser()
+
+            profile.saving = false
+            profile.fail = false
+            profile.success = false
+            profile.timezoneById = Timezones.timezoneById
+            profile.toggleOffFunctions = {}
+
+            profile.resetAllData = () => {
+                angular.copy(profile.userSecurityQuestions, profile.tempUserSecurityQuestions)
+                angular.copy(profile.user, profile.tempUser)
+            }
+
+            profile.toggleAllOff = () => {
+                angular.forEach(profile.toggleOffFunctions, function(toggleOff) {
+                    toggleOff()
+                })
+                profile.resetAllData()
+            }
+
+            profile.pushToggleOff = (toggleOffObject) => {
+                if (!profile.toggleOffFunctions) {
+                    profile.toggleOffFunctions = {}
+                }
+                profile.toggleOffFunctions[toggleOffObject.name] = toggleOffObject.function
+            }
+
+            profile.resetPasswordFields = () => {
+                profile.userPasswordAccount = {
+                    currentPassword: '',
+                    password: ''
+                }
+                profile.passwordRe = ''
+            }
+
+            profile.checkIfRepeatedSecurityAnswer = (securityQuestions, formObject) => {
+                securityQuestions.forEach((secQuestion, i) => {
+                    let securityAnswerRepeatedIndex = _.findIndex(securityQuestions, (secQuestionToCompareTo, z) => {
+                        return z !== i && secQuestion.answer && secQuestionToCompareTo.answer && secQuestion.answer.toUpperCase() === secQuestionToCompareTo.answer.toUpperCase()
+                    })
+                    if (securityAnswerRepeatedIndex > -1) {
+                        if (formObject['answer' + securityAnswerRepeatedIndex]) {
+                            formObject['answer' + securityAnswerRepeatedIndex].$setValidity('securityAnswerRepeated', false)
                         }
-                        else {
-                            if (formObject['answer' + i]) {
-                                formObject['answer'+i].$setValidity('securityAnswerRepeated', true);
-                            }
+                        if (formObject['answer' + i]) {
+                            formObject['answer' + i].$setValidity('securityAnswerRepeated', false)
                         }
-                    });
-                };
-
-                userProfile.resetChallengeQuestion = function(index) {
-                    userProfile.resetTempObject(userProfile.userSecurityQuestions.questions[index], userProfile.tempUserSecurityQuestions[index]);
-                };
-
-                userProfile.pushToggleOff = function(toggleOffObject) {
-                    if (!userProfile.toggleOffFunctions) {
-                        userProfile.toggleOffFunctions = {};
-                    }
-                    userProfile.toggleOffFunctions[toggleOffObject.name] = toggleOffObject.function;
-                };
-
-                userProfile.updatePerson = function(section, toggleOff) {
-                    if (section) {
-                        userProfile[section] = {
-                            submitting:true
-                        };
-                    }
-                    if (!userProfile.userCountry) {
-                        userProfile.tempUser.addresses[0].country = userProfile.user.addresses[0].country;
                     }
                     else {
-                        userProfile.tempUser.addresses[0].country = userProfile.userCountry.originalObject.code;
+                        if (formObject['answer' + i]) {
+                            formObject['answer' + i].$setValidity('securityAnswerRepeated', true)
+                        }
                     }
+                })
+            }
 
-                    // [7/20/2016] Note: Can't pass in 'activatedDate' anymore when updating a person
-                    delete userProfile.tempUser.activatedDate
-
-                    API.cui.updatePerson({personId: userId, data:userProfile.tempUser})
-                    .done(() => {
-                        angular.copy(userProfile.tempUser, userProfile.user);
-                        LocaleService.setLocaleByDisplayName(appConfig.languages[userProfile.user.language])
-                        if (toggleOff) {
-                            toggleOff();
-                        }
-                    })
-                    .fail((err) => {
-                        console.error('Failed to update user profile:', err)
-                        if (section) {
-                            userProfile[section].error = true;
-                        }
-                    })
-                    .always(() => {
-                        if (section) {
-                            userProfile[section].submitting = false;
-                        }
-                        $scope.$digest()
-                    })
-                };
-
-                userProfile.updatePassword = function(section, toggleOff) {
-                    if (section) {
-                        userProfile[section] = { submitting:true };
-                    }
-
-                    API.cui.updatePersonPassword({personId: userId, data: self.getPersonPasswordAccount(userProfile)})
-                    .always(() => {
-                        if (section) {
-                            userProfile[section].submitting = false;
-                        }
-                    })
-                    .done(() => {
-                        if (toggleOff) {
-                            toggleOff();
-                        }
-                        userProfile.passwordUpdateSuccess = true;
-                        $timeout(() => {
-                            userProfile.passwordUpdateSuccess = false;
-                        }, 5000);
-                        userProfile.resetPasswordFields();
-                        $scope.$digest();
-                    })
-                    .fail((err) => {
-                        console.error('Error updating password', err);
-                        if (section) {
-                            userProfile[section].error = true;
-                        }
-                        $scope.$digest();
-                    })
-                };
-
-                userProfile.saveChallengeQuestions = function(section, toggleOff) {
-                    if(section) {
-                        userProfile[section] = {
-                        submitting:true
-                    };
-
-                    userProfile.userSecurityQuestions.questions = angular.copy(userProfile.tempUserSecurityQuestions);
-                    self.selectTextsForQuestions(userProfile);
-
-                    API.cui.updateSecurityQuestionAccount({personId: userId,
-                        data: {
-                            version: '1',
-                            id: userId,
-                            questions: userProfile.userSecurityQuestions.questions
-                        }
-                    })
-                    .then(function(res) {
-                        if (section) {
-                            userProfile[section].submitting = false;
-                        }
-                        if (toggleOff) {
-                            toggleOff();
-                        }
-                        $scope.$digest();
-                    })
-                    .fail(function(err) {
-                        console.log(err);
-                        if(section) {
-                            userProfile[section].submitting = false;
-                        }
-                        if(section) {
-                            userProfile[section].error = true;
-                        }
-                        $scope.$digest();
-                    });
+            profile.updatePerson = (section, toggleOff) => {
+                if (section) {
+                    profile[section] = { submitting: true }
                 }
-            };
+
+                if (!profile.userCountry) {
+                    profile.tempUser.addresses[0].country = profile.user.addresses[0].country;
+                } else {
+                    profile.tempUser.addresses[0].country = profile.userCountry.originalObject.code;
+                }
+
+                // [7/20/2016] Note: Can't pass in 'activatedDate' anymore when updating a person
+                delete profile.tempUser.activatedDate
+
+                API.cui.updatePerson({personId: userId, data:profile.tempUser})
+                .always(() => {
+                    if (section) {
+                        profile[section].submitting = false;
+                    }
+                    $scope.$digest()
+                })
+                .done(() => {
+                    angular.copy(profile.tempUser, profile.user)
+                    LocaleService.setLocaleByDisplayName(appConfig.languages[profile.user.language])
+                    if (toggleOff) {
+                        toggleOff();
+                    }
+                })
+                .fail((err) => {
+                    console.error('Failed to update user profile:', err)
+                    if (section) {
+                        profile[section].error = true;
+                    }
+                })
+            }
+
+            profile.updatePassword = function(section, toggleOff) {
+                if (section) profile[section] = { submitting: true }
+
+                API.cui.updatePersonPassword({ 
+                    personId: userId, 
+                    data: UserProfile.buildPersonPasswordAccount(profile.user, profile.userPasswordAccount, profile.organization) 
+                })
+                .always(() => {
+                    if (section) profile[section].submitting = false
+                })
+                .done(() => {
+                    if (toggleOff) toggleOff()
+                    profile.passwordUpdateSuccess = true
+                    $timeout(() => profile.passwordUpdateSuccess = false, 5000)
+                    profile.resetPasswordFields()
+                    $scope.$digest()
+                })
+                .fail(err => {
+                    console.error('Error updating password', err)
+                    if (section) profile[section].error = true
+                    $scope.$digest()
+                })
+            }
+
+            profile.saveChallengeQuestions = (section, toggleOff) => {
+                if (section) {
+                    profile[section] = { submitting: true }
+                }
+                profile.userSecurityQuestions = angular.copy(profile.tempUserSecurityQuestions)
+
+                API.cui.updateSecurityQuestionAccount({
+                    personId: userId,
+                    data: {
+                        version: '1',
+                        id: userId,
+                        questions: profile.userSecurityQuestions.questions
+                    }
+                })
+                .always(() => {
+                    if (section) {
+                        profile[section].submitting = false;
+                    }
+                })
+                .done(() => {
+                    if (toggleOff) {
+                        toggleOff();
+                    }
+                    profile.challengeQuestionsTexts = UserProfile.selectTextsForQuestions(profile)
+                    $scope.$digest()
+                })
+                .fail(err => {
+                    console.error('Error updating security questions', err)
+                    if (section) {
+                        profile[section].error = true;
+                    }
+                    $scope.$digest()
+                })
+            }
+
+            profile.validatePassword = (password, formObject, input) => {
+
+                const validSwitch = (input, isValidBoolean) => {
+                    switch (input) {
+                        case 'newPassword':
+                            profile.validNewPassword = isValidBoolean
+                        case 'newPasswordRe':
+                            profile.validNewPasswordRe = isValidBoolean
+                    }
+                }
+
+                const validateData = {
+                    userId: userId,
+                    organizationId: profile.user.organization.id,
+                    password: password,
+                    operations: ['PASSWORD_SPECIFY']
+                }
+
+                API.cui.validatePassword({data: validateData})
+                .then(res => {
+                    let validPasswordHistory = false
+
+                    res.forEach(rule => {
+                        if (rule.type === 'HISTORY' && rule.isPassed) {
+                            validPasswordHistory = true
+                            return
+                        }
+                    })
+
+                    if (validPasswordHistory) {
+                        validSwitch(input, true)
+                        formObject[input].$setValidity(input, true)
+                        $scope.$digest()
+                    }
+                    else {
+                        validSwitch(input, false)
+                        formObject[input].$setValidity(input, false)
+                        $scope.$digest()
+                    }
+                })
+            }
         }
-    };
+    }
 
-    return self;
-
-});
+    return UserProfile
+})