+++ app/modules/user/profile/user-profile.ctrl.js
@@ -1,16 +1,16 @@
 angular.module('user')
-.controller('userProfileCtrl', function(Loader, User, UserProfileV2, $scope) {
+.controller('userProfileCtrl', function(Loader, User, UserProfile, $scope) {
 
     const userProfile = this
     const scopeName = 'userProfile.'
 
     /* -------------------------------------------- ON LOAD START --------------------------------------------- */
 
-    UserProfileV2.injectUI(userProfile, $scope, User.user.id)
+    UserProfile.injectUI(userProfile, $scope, User.user.id)
 
     Loader.onFor(scopeName + 'initProfile')
 
-    UserProfileV2.initUserProfile(User.user.id, User.user.organization.id)
+    UserProfile.initUserProfile(User.user.id, User.user.organization.id)
     .then(res => {
         angular.merge(userProfile, res)
         Loader.offFor(scopeName + 'initProfile')